Added footer and reorganized page files into separate folder
