+++ src/components/footer/Footer.jsx
@@ -1,0 +1,9 @@
+import "../../styles/footer.css";
+
+export default function Footer(){
+    return (
+        <footer>
+            <h3 className="copyright">Copyright © 2024 Weston Collegiate Institute</h3>
+        </footer>
+    );
+};
